requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Solana dependencies
solana==0.30.2
solders==0.18.1

# AI/ML dependencies
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2
tensorflow==2.16.1
prophet==1.1.5
ta==0.10.2  # Technical Analysis library

# Data processing
pydantic==2.5.0
python-multipart==0.0.6

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0

# Utilities
schedule==1.2.0
pytz==2023.3